+++ opengever/testing/builders/sql.py
@@ -43,7 +43,7 @@ from opengever.meeting.proposal import Proposal
 from opengever.ogds.base.interfaces import IAdminUnitConfiguration
 from opengever.ogds.base.utils import get_ou_selector
 from opengever.ogds.models.admin_unit import AdminUnit
-from opengever.ogds.models.tests.builders import OrgUnitBuilder
+from opengever.ogds.models.org_unit import OrgUnit
 from opengever.ogds.models.tests.builders import UserBuilder
 from opengever.task.reminder import TASK_REMINDER_SAME_DAY
 from opengever.testing.builders.base import TEST_USER_ID
@@ -152,40 +152,95 @@ class AdminUnitBuilder(SqlObjectBuilder):
 builder_registry.register('admin_unit', AdminUnitBuilder)
 
 
-class PloneOrgUnitBuilder(OrgUnitBuilder):
-    """Add plone specific functionality to opengever.ogds.models
-    OrgUnitBuilder.
+class OrgUnitBuilder(SqlObjectBuilder):
 
-    """
+    mapped_class = OrgUnit
+    id_argument_name = 'unit_id'
+
     def __init__(self, session):
-        super(PloneOrgUnitBuilder, self).__init__(session)
+        super(OrgUnitBuilder, self).__init__(session)
+        self.arguments[self.id_argument_name] = u'rr'
+        self.arguments['users_group_id'] = 'foo'
+        self.arguments['inbox_group_id'] = 'bar'
+        self._with_inbox_group = False
+        self._with_users_group = False
+        self._inbox_users = set()
+        self._group_users = set()
         self._as_current_org_unit = False
+
+    def before_create(self):
+        self._assemble_groups()
 
     def after_create(self, obj):
         if self._as_current_org_unit:
             get_ou_selector().set_current_unit(obj.id())
         return obj
 
+    def with_default_groups(self):
+        self.with_inbox_group()
+        self.with_users_group()
+        return self
+
+    def with_inbox_group(self):
+        self._with_inbox_group = True
+        return self
+
+    def with_users_group(self):
+        self._with_users_group = True
+        return self
+
     def as_current_org_unit(self):
         self._as_current_org_unit = True
         return self
+
+    def assign_users(self, users, to_users=True, to_inbox=True):
+        if to_users:
+            self.with_users_group()
+            self._group_users.update(users)
+
+        if to_inbox:
+            self.with_inbox_group()
+            self._inbox_users.update(users)
+        return self
+
+    def _assemble_groups(self):
+        if self._with_users_group or self._with_inbox_group:
+            unit_id = self.arguments.get(self.id_argument_name)
+
+        if self._with_users_group:
+            users_group_id = "{0}_users".format(unit_id)
+            users_group_title = '{0} Users Group'.format(unit_id)
+            self._create_users_group(users_group_id, users_group_title)
+
+        if self._with_inbox_group:
+            users_inbox_id = "{0}_inbox_users".format(unit_id)
+            users_inbox_title = '{0} Inbox Users Group'.format(unit_id)
+            self._create_inbox_group(users_inbox_id, users_inbox_title)
 
     def _create_users_group(self, users_group_id, users_group_title=None):
         create(Builder('group')
                .having(title=users_group_title)
                .with_groupid(users_group_id)
                .with_members(api.user.get(TEST_USER_ID)))
-        super(PloneOrgUnitBuilder, self)._create_users_group(users_group_id)
+        users_group = create(Builder('ogds_group')
+                             .having(groupid=users_group_id,
+                                     title=users_group_title,
+                                     users=list(self._group_users)))
+        self.arguments['users_group'] = users_group
 
     def _create_inbox_group(self, users_inbox_id, users_inbox_title=None):
         create(Builder('group')
                .having(title=users_inbox_title)
                .with_groupid(users_inbox_id)
                .with_members(api.user.get(TEST_USER_ID)))
-        super(PloneOrgUnitBuilder, self)._create_inbox_group(users_inbox_id)
+        inbox_group = create(Builder('ogds_group')
+                             .having(groupid=users_inbox_id,
+                                     title=users_inbox_title,
+                                     users=list(self._inbox_users)))
+        self.arguments['inbox_group'] = inbox_group
 
 
-builder_registry.register('org_unit', PloneOrgUnitBuilder, force=True)
+builder_registry.register('org_unit', OrgUnitBuilder)
 
 
 class PloneOGDSUserBuilder(UserBuilder):
